Bug fixes in CUDD 2.4.2.

+++ src/bdd/cudd/cuddInteract.c
@@ -193,10 +193,10 @@ cuddTestInteract(
   int  x,
   int  y)
 {
-    long posn, word, bit, result;
+    int posn, word, bit, result;
 
     if (x > y) {
-        long tmp = x;
+        int tmp = x;
         x = y;
         y = tmp;
     }
